Move generic TLS tests to test_tls.cpp
Leaves unit_tls.cpp for the handshake level tests.

Add some basic tests of the string<->enum conversions in tls_algos.h
